% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepper.R, R/transformer.R
\name{add_node}
\alias{add_node}
\alias{add_node}
\title{add a pre-processing stage}
\usage{
add_node(x, ...)

add_node(x, ...)
}
\arguments{
\item{x}{the processing pipeline}

\item{x}{the processing pipeline}
}
\description{
add a pre-processing stage

add a pre-processing stage
}
