% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bada.R
\name{bada}
\alias{bada}
\title{Barycentric Discriminant Analysis}
\usage{
bada(Y, X, S, ncomp = length(levels(as.factor(Y))) - 1, center = TRUE,
  scale = FALSE, ...)
}
\arguments{
\item{Y}{dependent \code{factor} variable. The categories of the observations.}

\item{X}{the data matrix with \code{nrow} equal to \code{length} of \code{Y}}

\item{S}{an integer \code{vector} or \code{factor} of subject ids which is the same length as \code{Y}.}

\item{ncomp}{number of common components to estimate}

\item{center}{whether to center the variables}

\item{scale}{whether to divide the variables by the standard deviation.}

\item{...}{arguments to pass through}
}
\description{
A technique whose goal is to assign observations to some predetermined categories a  linear combination of the
variables of a data matrix
}
\examples{

X <- matrix(rnorm(100*1000), 100, 1000)
Y <- factor(rep(letters[1:4], 25))
S <- factor(rep(1:5, each=20))

bres <- bada(Y, X, S, ncomp=3)
}
\references{
Abdi, H., Williams, L. J., & Béra, M. (2017). Barycentric discriminant analysis. \emph{Encyclopedia of Social Network Analysis and Mining}, 1-20.
}
