% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{block_apply}
\alias{block_apply}
\title{block_apply
apply a function to each block of a multi-block data structure.}
\usage{
block_apply(x, f, ...)
}
\arguments{
\item{x}{the multi-block data}

\item{f}{the function to apply}

\item{...}{extra args}
}
\description{
block_apply
apply a function to each block of a multi-block data structure.
}
