% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{block_index_list}
\alias{block_index_list}
\title{block_index_list}
\usage{
block_index_list(object)
}
\arguments{
\item{object}{the object}
}
\description{
extract the list of indices associated with each block in a multi-block object
}
