% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{block_lengths}
\alias{block_lengths}
\title{block_lengths}
\usage{
block_lengths(object)
}
\arguments{
\item{object}{the object}
}
\description{
extract the vector of lengths of each block in a multi-block object
}
