% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_matrix.R
\name{block_matrix_list}
\alias{block_matrix_list}
\title{block_matrix_list}
\usage{
block_matrix_list(Xs)
}
\arguments{
\item{Xs}{a list of k matrices each with with N_k rows and M_k columns}
}
\description{
block_matrix_list
}
