% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{block_project}
\alias{block_project}
\title{project a single 'block' of data onto the subspace}
\usage{
block_project(x, newdata, block, ...)
}
\arguments{
\item{x}{the model fit}

\item{newdata}{a matrix or vector of new obervations(s)}

\item{block}{the block id to select in the block projection matrix}

\item{...}{extra args}
}
\description{
project a single 'block' of data onto the subspace
}
