% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_projector.R
\name{block_projector}
\alias{block_projector}
\title{#' projector_list
projector_list <- function(Xs) {
  assert_that(all(sapply(Xs, function(x) inherits(x, "projector"))))
  structure(Xs,
            block_indices=block_indices(Xs),
            class=c("projector_list", "list"))
}}
\usage{
block_projector(Xs)
}
\arguments{
\item{Xs}{a list of \code{projector} objects.}
}
\description{
#' @export
project.projector_list(x, newdata, block_index=NULL) {
  if (!is.null(block_index)) {
    assert_that(length(block_index) == 1 || length(block_index) == length(x))
  }
  if (is.null(block_index)) {
    
  }
}
block_projector
}
\details{
a blockwise concatenation of a list of \code{projectors} each with the same row-dimension
}
