% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{bootstrap}
\alias{bootstrap}
\title{bootstrap a model}
\usage{
bootstrap(x, nboot, ...)
}
\arguments{
\item{x}{the model fit}

\item{nboot}{the number of bootstrap resamples}

\item{...}{extra args}
}
\description{
bootstrap a model
}
