% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{classifier}
\alias{classifier}
\title{turn a model object into a classifier}
\usage{
classifier(x, ...)
}
\arguments{
\item{x}{the fitted object}
}
\description{
turn a model object into a classifier
}
