% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asca.R
\name{collapse}
\alias{collapse}
\title{Column-wise average a matrix of variables, collapsing over some set of factors}
\usage{
collapse(form, X, design)
}
\arguments{
\item{form}{the formula defining the model to collapse over}

\item{X}{the matrix to collpased over rows}

\item{design}{the \code{data.frame} containing the variables reference in \code{form}}
}
\description{
Column-wise average a matrix of variables, collapsing over some set of factors
}
\examples{
X <- matrix(rnorm(20*10), 20, 10)
des <- data.frame(a=rep(letters[1:4], 5), b=factor(rep(1:5, each=4)))
xcoll <- collapse(~ a+b, X, design=des)
}
