% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{contributions}
\alias{contributions}
\title{contributions}
\usage{
contributions(x, ...)
}
\arguments{
\item{x}{the model fit}

\item{...}{extra args}
}
\description{
compute the contributions (of observations, variables, tables) to a model.
}

