% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{cross_validate}
\alias{cross_validate}
\title{cross_validate a model}
\usage{
cross_validate(x, ...)
}
\arguments{
\item{x}{the model fit}

\item{...}{extra args}
}
\description{
cross_validate a model
}
