% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepper.R, R/transformer.R
\name{dim_reduce}
\alias{dim_reduce}
\alias{dim_reduce}
\title{dimension reduction as a pre-processing stage}
\usage{
dim_reduce(preproc = prepper(), method = pca, ...)

dim_reduce(preproc = prepper(), method = pca, ...)
}
\arguments{
\item{preproc}{the pipeline}

\item{method}{the dimension reduction method (e.g. pca)}

\item{preproc}{the pipeline}

\item{method}{the dimension reduction method (e.g. pca)}
}
\description{
dimension reduction as a pre-processing stage

dimension reduction as a pre-processing stage
}
