% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpp.R
\name{embed_lpp}
\alias{embed_lpp}
\title{locality preserving projections}
\usage{
embed_lpp(X, W, ndim = 2)
}
\arguments{
\item{X}{the data matrix}

\item{W}{the affinity matrix}

\item{ndim}{number of dimensions}
}
\description{
locality preserving projections
}

