% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepper.R
\name{fresh}
\alias{fresh}
\title{get a fresh pre-processing node cleared of any cached data}
\usage{
fresh(x, ...)
}
\arguments{
\item{x}{the processing pipeline}
}
\description{
get a fresh pre-processing node cleared of any cached data
}
