% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpca.R
\name{genpca}
\alias{genpca}
\title{Generalized Principal Components Analysis}
\usage{
genpca(
  X,
  A = NULL,
  M = NULL,
  ncomp = min(dim(X)),
  preproc = center(),
  deflation = FALSE,
  svd_init = TRUE,
  threshold = 1e-06,
  use_cpp = TRUE,
  maxeig = 800
)
}
\arguments{
\item{A}{the column constraints. Can be a \code{vector}, symmetric \code{matrix}, or symmetric sparse matrix with \code{ncol(X)} rows and columns.}

\item{M}{the row constraints. Can be a \code{vector}, symmetric \code{matrix}, or symmetric sparse matrix with \code{nrow(X)} rows and columns.}
}
\description{
Compute a PCA in a inner-product space defined by row and coulmn constraint matrices.
}
\examples{
X <- matrix(rnorm(20*10), 20, 10)
res <- genpca(X, ncomp=5, preproc=center())
scores(res)
}
\references{
Abdi, H. (2007). Singular value decomposition (SVD) and generalized singular value decomposition. \emph{Encyclopedia of measurement and statistics}, 907-912.

Allen, G. I., Grosenick, L., & Taylor, J. (2014). A generalized least-square matrix decomposition. \emph{Journal of the American Statistical Association}, 109(505), 145-159.
}
