% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_matrix.R
\name{get_block}
\alias{get_block}
\title{get_block}
\usage{
get_block(x, i, ...)
}
\arguments{
\item{x}{the object to retrive the block of variables from.}

\item{i}{the index of the block.}
}
\description{
get_block
}

