% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_means.R, R/utils.R
\name{group_means}
\alias{group_means}
\title{group_means}
\usage{
group_means(Y, X)

group_means(Y, X)
}
\arguments{
\item{Y}{\code{factor} variable defining the groups}

\item{X}{\code{matrix} defining the matrix data to be group-wise averaged}

\item{Y}{\code{factor} variable defining the groups}

\item{X}{\code{matrix} defining the matrix data to be group-wise averaged}
}
\description{
group_means
}

