% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subpca.R
\name{hpca}
\alias{hpca}
\title{hclust_pca}
\usage{
hpca(X, hclus, cuts, est_method = c("fixed", "gcv", "shrink"),
  ncomp = rep(1, length(cuts) + 1), center = TRUE, scale = FALSE,
  shrink_method = "GSURE")
}
\arguments{
\item{X}{the data matrix}

\item{hclus}{a hierarchical clustering instance with as many objects as there are rows in \code{X}}

\item{cuts}{desired number of clusters at each level of the hierarchy (must be increasing)}

\item{ncomp}{the number of components to estimate at each level}
}
\description{
A type of pca that uses a hierarchical clustering to define a set of nested regions for pca compression.
}
\examples{

grid <- expand.grid(1:10, 1:10)
X <- matrix(rnorm(100*50), 100, 50)
cuts <- c(4, 8, 16)
hclus <- hclust(dist(grid))
hres1 <- hpca(X, hclus, cuts, est_method="fixed", ncomp=c(4,1,1,1))
ncomp(hres1) == (sum(cuts) +4)

hres2 <- hpca(X, hclus, cuts, est_method="shrink")
hres3 <- hpca(X, hclus, cuts, est_method="gcv")

}
