% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{loadings}
\alias{loadings}
\title{loadings}
\usage{
loadings(x, ...)
}
\arguments{
\item{x}{the model fit}

\item{...}{extra args}
}
\description{
extract the loadings matrix (the variable coefficients) for a multivariate model.
}
