% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapca.R
\name{metapca}
\alias{metapca}
\title{metapca}
\usage{
metapca(fits, ncomp = 2, scale_inner = FALSE)
}
\arguments{
\item{scale_inner}{standardize first level components

X1 <- matrix(rnorm(20*10), 20, 10)
X2 <- matrix(rnorm(20*20), 20, 20)
X3 <- matrix(rnorm(20*30), 20, 30)

pc1 <- pca(X1, ncomp=10, preproc=center())
pc2 <- pca(X2, ncomp=10, preproc=center())
pc3 <- pca(X3, ncomp=10, preproc=center())

Xg <- cbind(X1,X2,X3)

fits <- list(pc1,pc2,pc3)
pfit <- metapca(fits, ncomp=15, scale_inner=TRUE)
ncol(scores(pfit)) == 10}
}
\description{
metapca
}
