% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{mfa}
\alias{mfa}
\title{multiple factor analysis}
\usage{
mfa(X, ncomp = 2, center = TRUE, scale = FALSE, normalization = c("MFA",
  "RV", "None", "RV-MFA", "custom"), A = NULL)
}
\arguments{
\item{X}{a \code{block_matrix} or a \code{block_projection_matrix} object}

\item{ncomp}{the number of components to estimate}

\item{center}{whether to mean center the variables}

\item{scale}{whether to standardize the variables}

\item{normalization}{the normalization method: MFA, RV, RV-MFA, or None (see details).}
}
\description{
mfa
}
\examples{

X <- block_matrix(replicate(5, { matrix(rnorm(10*10), 10, 10) }, simplify=FALSE))
res <- mfa(X, ncomp=3, normalization="MFA")
p <- partial_scores(res, 1)
stopifnot(ncol(scores(res)) == 3)
}
