% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmpca.R
\name{mmpca}
\alias{mmpca}
\title{mmpca}
\usage{
mmpca(X, Y, ncomp = min(dim(X)), center = TRUE, scale = FALSE, knn = 1,
  sigma = 0.7)
}
\arguments{
\item{X}{the data matrix with \code{nrow} equal to \code{length} of \code{Y}}

\item{Y}{dependent \code{factor} variable. The categories of the observations.}

\item{ncomp}{number of common components to estimate}

\item{center}{whether to center the variables}

\item{scale}{whether to divide the variables by the standard deviation.}

\item{knn}{the number of nearest neighbors used to form margin}
}
\description{
mmpca
}
