% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asca.R
\name{muasca}
\alias{muasca}
\title{muasca}
\usage{
muasca(formula, Xlist, ncomp = 2, design, scale = FALSE, A = NULL)
}
\arguments{
\item{formula}{a formula specifying the ANOVA design}

\item{Xlist}{a \code{list} of X matrices, one per subject.}

\item{design}{a \code{data.frame} providing the variables provided in \code{formula} argument.}

\item{scale}{whether to scale the variables}
}
\description{
muasca
}
\examples{

f1 <- factor(rep(letters[1:4], 25))
f2 <- factor(rep(1:4, each=25))

des <- data.frame(f1=f1,f2=f2)

Xlist <- replicate(5, matrix(rnorm(100*100), 100, 100), simplify=FALSE)

res <- muasca(~ f1*f2, Xlist, ncomp=3,design=des)
}
