% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asca.R
\name{musu_asca}
\alias{musu_asca}
\title{plscorr_aov}
\usage{
musu_asca(Xlist, formula, design, svd.method = "base")
}
\arguments{
\item{Xlist}{the data matrices, 1 per subject}

\item{formula}{a formula specifying the ANOVA design}

\item{design}{a \code{data.frame} providing the variables provided in \code{formula} argument.}

\item{svd.method}{the svd method to use.}
}
\description{
plscorr_aov
}

