% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musubada.R
\name{musubada}
\alias{musubada}
\title{musubada}
\usage{
musubada(Y, Xlist, ncomp = 2, center = TRUE, scale = FALSE,
  svd.method = "fast", normalization = c("MFA", "RV", "None", "DCor"))
}
\arguments{
\item{Y}{dependent \code{factor} variable. If All X matrices have same number of rows, Y can be a single factor.
If there are a different nume rof rows (e..g different numbers of replications per subject), Y can be a list of factors.}

\item{Xlist}{a list of X matrices, one per subject.}
}
\description{
musubada
}

