% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{nblocks}
\alias{nblocks}
\title{nblocks
extract the number of blocks in a mutli-block data structure or model}
\usage{
nblocks(x)
}
\arguments{
\item{x}{the object}
}
\description{
nblocks
extract the number of blocks in a mutli-block data structure or model
}
