% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{ncomp}
\alias{ncomp}
\title{ncomp get the number of components in the estimated model}
\usage{
ncomp(x)
}
\arguments{
\item{x}{the model fit}
}
\description{
ncomp get the number of components in the estimated model
}
