% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroca.R
\docType{package}
\name{neuroca}
\alias{neuroca-package}
\alias{neuroca}
\title{neuroca - multivariate component analysis methods for neuroimaging data}
\description{
neuroca contains methods for multivariate analysis of neuroimaging data
}
\author{
Bradley Buchsbaum <brad.buchsbaum@gmail.com>
}
