% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuropls.R
\docType{package}
\name{neuropls}
\alias{neuropls}
\alias{neuropls-package}
\title{neuropls - a multivariate analysis of neuroimaging data}
\description{
neuropls - a multivariate analysis of neuroimaging data
}
\author{
Bradley Buchsbaum <brad.buchsbaum@gmail.com>
}
