% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{partial_project}
\alias{partial_project}
\title{project a selected subset of indices onto the subspace}
\usage{
partial_project(x, newdata, colind)
}
\arguments{
\item{x}{the model fit}

\item{newdata}{a matrix or vector of new obervations(s)}

\item{the}{column indices to select in the projection matrix}
}
\description{
project a selected subset of indices onto the subspace
}
