% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{partial_projector}
\alias{partial_projector}
\title{partial_projector}
\usage{
partial_projector(x, colind, ...)
}
\arguments{
\item{x}{the model fit}

\item{colind}{a vector of unique subindices}

\item{...}{extra args}
}
\description{
extract a partial projector from a fitted model
}
