% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{partial_scores}
\alias{partial_scores}
\title{partial_scores}
\usage{
partial_scores(x, ...)
}
\arguments{
\item{x}{the model fit}
}
\description{
compute the partial scores from a multivariate model using a subset of the input
}
