% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{principal components analysis}
\usage{
pca(X, ncomp = min(dim(X)), preproc = center(), ...)
}
\description{
principal components analysis
}
\examples{

X <- matrix(rnorm(10*20), 10, 20)
res <- pca(X, ncomp=10, preproc=center())
}
