% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_config.R
\name{PLS_Config}
\alias{PLS_Config}
\title{PLS_Config
Configuration information for PLS and related mutivariate models}
\usage{
PLS_Config(base_path = "", design_list, data_format = "image",
  folding_dimension = NULL, blocking_dimension = NULL, block_ids = NULL,
  model_name = "Mean_Centered_PLS", model_control = list())
}
\arguments{
\item{folding_dimension}{name of dimension will be folded into the columns of the X matrix.}

\item{blocking_dimension}{this dimension will be treated as experimental units or blocks (e.g. subject, session)}

\item{block_ids}{}

\item{model_name}{the name of the statistical model to be applied to dataset.}

\item{model_control}{a list of control parameters for the model algorithm.}

\item{data_files}{the names of the image data files}

\item{design}{the file or list of files containing the design matrix information}
}
\description{
PLS_Config
Configuration information for PLS and related mutivariate models
}

