% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{project}
\alias{project}
\title{project}
\usage{
project(x, newdata, ...)
}
\arguments{
\item{x}{the model fit}

\item{newdata}{a matrix or vector of new obervations(s)}

\item{...}{extra args}
}
\description{
project supplementary observations in to the subspace defined by the model
}
