% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{project_cols}
\alias{project_cols}
\title{project_cols}
\usage{
project_cols(x, newdata, ...)
}
\arguments{
\item{x}{the model fit}

\item{newdata}{a matrix or vector of new variables(s)}

\item{...}{extra args}
}
\description{
project supplementary variables in to the subspace defined by the model
}
