% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pseudo_pca}
\alias{pseudo_pca}
\title{pseudo_pca}
\usage{
pseudo_pca(u, v, d, pre_process = function(x, subind) x,
  reverse_pre_process = function(x, subind) x, rnames = NULL)
}
\arguments{
\item{u}{the row weights}

\item{v}{the column weights}

\item{d}{the singular values}

\item{pre_process}{the function to pre_process the data}

\item{reverse_pre_process}{the function to reverse pre_process the data}

\item{row}{names of observations}
}
\description{
pseudo_pca
}
