% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pseudo_svd}
\alias{pseudo_svd}
\title{pseudo_svd}
\usage{
pseudo_svd(u, v, d, rnames = NULL)
}
\arguments{
\item{u}{the row weights}

\item{v}{the column weights}

\item{d}{the singular values}

\item{rnames}{row names of observations}
}
\description{
pseudo_svd
}
