% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{reconstruct}
\alias{reconstruct}
\title{reconstruct the data with some number of components}
\usage{
reconstruct(x, newdata, comp, ...)
}
\arguments{
\item{x}{the model fit}

\item{newdata}{newdata to be inverse projected (optional)}

\item{comp}{the components to use}

\item{...}{extra args}
}
\description{
reconstruct the data with some number of components
}
