% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{reduce_rank}
\alias{reduce_rank}
\title{reduce_rank}
\usage{
reduce_rank(x, k, ...)
}
\arguments{
\item{x}{the data matrix with m columns}

\item{k}{the desired number of components}
}
\description{
reduce the rank of a data set to from m variables to k components
}
