% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{reproducibility}
\alias{reproducibility}
\title{reproducibility}
\usage{
reproducibility(x, folds, metric, ...)
}
\arguments{
\item{x}{the model fit}
}
\description{
compute a measure of the reproducibility of a model under replication.
}

