% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{resample}
\alias{resample}
\title{resample data from a model fit}
\usage{
resample(x, ...)
}
\arguments{
\item{x}{the model fit}

\item{...}{extra args}
}
\description{
resample data from a model fit
}
