% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{rotate}
\alias{rotate}
\title{apply a rotation matrix to a solution}
\usage{
rotate(x, rot, ...)
}
\arguments{
\item{x}{the object to rotate}

\item{rot}{the rotation matrix to apply}

\item{...}{extra arguments}
}
\description{
apply a rotation matrix to a solution
}
