% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scada.R
\name{scada}
\alias{scada}
\title{scada}
\usage{
scada(
  Y,
  Xlist,
  ncomp = 2,
  center = TRUE,
  scale = FALSE,
  rank_k = NULL,
  type = c("sca-p", "sca-pf2", "sca-ind", "sca-ecp")
)
}
\arguments{
\item{Y}{dependent \code{factor} variable. If All X matrices have same number of rows, Y can be a single factor.
If there are a different number of rows (e.g. different numbers of replications per subject), Y can be a list of factors.}

\item{Xlist}{a list of X matrices, one per subject.}

\item{ncomp}{number of common components to estimate}

\item{center}{whether to center the variables}

\item{scale}{whether to scale the variables by 1/sd}

\item{rank_k}{reduce data to k components per block via pca}

\item{normalization}{the type of normalization}
}
\description{
simulataneous component discriminant analysis
}
