% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{shrink_pca}
\alias{shrink_pca}
\title{shrink_pca}
\usage{
shrink_pca(X, center = TRUE, scale = FALSE, method = c("GSURE", "QUT",
  "SURE"), ...)
}
\description{
adaptive shrinakge pca from the \code{denoiseR} package
}
