% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{spatial_constraints}
\alias{spatial_constraints}
\title{compute a spatial contraint matrix that can span across blocks}
\usage{
spatial_constraints(coords, nblocks = 1, feature_scores = rep(1,
  nrow(coords) * nblocks), sigma = 5, sigma_between = 5,
  shrinkage_factor = 0.1)
}
\arguments{
\item{feature_scores}{a vector of feature weights used to weight each individual feature}

\item{sigma}{the within block smoothing radius}

\item{sigma_between}{the across block smoothing radius}

\item{shrinkage_factor}{the amount to weight the across block smoothing matrix (heiger leads to more shrinkae)}

\item{a}{matrix of coordinates for one block that must replicate exactly across \code{nblocks} blocks}
}
\description{
compute a spatial contraint matrix that can span across blocks
}
