% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd_wrapper.R
\name{svd_wrapper}
\alias{svd_wrapper}
\title{svd_wrapper}
\usage{
svd_wrapper(
  X,
  ncomp = min(dim(X)),
  method = c("base", "fast", "irlba", "propack", "rsvd", "svds"),
  tol = .Machine$double.eps,
  ...
)
}
\arguments{
\item{X}{the \code{matrix}}

\item{ncomp}{number of components to estimate}

\item{method}{the svd method to use. One of: 'base', 'fast', 'irlba', 'propack'}
}
\description{
svd_wrapper
}
