% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_ratio.R
\name{trace_ratio}
\alias{trace_ratio}
\title{trace_ratio optimization}
\usage{
trace_ratio(A, B, ncomp, eps = 1e-06, maxiter = 100)
}
\arguments{
\item{A}{the numerator matrix}

\item{B}{the denominator matrix}
}
\description{
trace_ratio optimization
}
\keyword{2012}
\keyword{Ngo}
\keyword{al}
\keyword{et}
