% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bada.R
\name{bada}
\alias{bada}
\title{Barycentric Discriminant Analysis}
\usage{
bada(Y, X, S = rep(1, nrow(X)), ncomp = length(levels(as.factor(Y))) - 1,
  preproc = neuroca::center(), A = NULL, M = NULL, ...)
}
\arguments{
\item{Y}{dependent \code{factor} variable. The categories of the observations.}

\item{X}{the data matrix with \code{nrow} equal to \code{length} of \code{Y}}

\item{S}{an integer \code{vector} or \code{factor} of subject ids which is the same length as \code{Y}.}

\item{ncomp}{number of components to estimate}

\item{preproc}{pre-processing function, defaults to \code{center}}

\item{A}{the column constraints}

\item{M}{the row constraints}

\item{...}{arguments to pass through}
}
\description{
A component technique that maximizes the between group variance over a set of variables.
}
\details{
The \code{S} argument can be used to model multi-level structure. If \code{S} is included, then pre-processing is applied 
separately to each unique value of S. This has the effect, for example when preproc = \code{center} of removing subject-specific 
means before computing the barycetners.
}
\examples{

X <- matrix(rnorm(100*1000), 100, 1000)
Y <- factor(rep(letters[1:4], length.out=100))
S <- factor(rep(1:10, each=10))

bres <- bada(Y, X, S, ncomp=3)

project(bres, X[S==1,])

## no strata
bres <- bada(Y, X, ncomp=3)



xbar <- matrix(rnorm(4*1000), 4, 1000)
Y <- factor(rep(letters[1:4], length.out=100))
X <- do.call(rbind, lapply(as.character(Y), function(y) { switch(y,
                               "a" = xbar[1,] + rnorm(1000)*5,
                               "b" = xbar[2,] + rnorm(1000)*5,
                               "c" = xbar[3,] + rnorm(1000)*5,
                               "d" = xbar[4,] + rnorm(1000)*5)
                               }
                               ))
                               
bres2 <- bada(Y, X, S, ncomp=3)


}
\references{
Abdi, H., Williams, L. J., & Béra, M. (2017). Barycentric discriminant analysis. \emph{Encyclopedia of Social Network Analysis and Mining}, 1-20.
}
