% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subpca.R
\name{block_pca}
\alias{block_pca}
\title{block_pca}
\usage{
block_pca(X, est_method = c("gcv", "shrink", "fixed"), ncomp = 2,
  min_k = 1, max_k = 3, center = TRUE, scale = FALSE,
  ncores = max(parallel::detectCores()/2, 1), ...)
}
\description{
block_pca
}
