% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asca.R
\name{collapse}
\alias{collapse}
\title{collapse}
\usage{
collapse(form, X, design)
}
\arguments{
\item{form}{the formula defining the model to collapse over}

\item{X}{the matrix to collpased ober rows}

\item{design}{the \code{data.frame} containing the variables reference in \code{form}}
}
\description{
collapse
}

