% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepper.R
\name{dim_reduce}
\alias{dim_reduce}
\title{dimension reduction as a pre-processing stage}
\usage{
dim_reduce(preproc = prepper(), method = pca, ...)
}
\arguments{
\item{preproc}{the pipeline}

\item{method}{the dimension reduction method (e.g. pca)}

\item{...}{args to be passed to the dimension reduction method}
}
\description{
dimension reduction as a pre-processing stage
}
