% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpca.R
\name{genpca}
\alias{genpca}
\title{genpca}
\usage{
genpca(X, A = rep(1, ncomp(X)), M = rep(1, nrow(X)), ncomp = min(dim(X)),
  center = TRUE, scale = FALSE)
}
\arguments{
\item{X}{the data matrix}

\item{A}{the column constraints. Can be a \code{vector}, \code{matrix}, or sparse matrix.}

\item{M}{the row constraints. Can be a \code{vector}, \code{matrix}, or sparse matrix.}

\item{ncomp}{the number of components to estimate}

\item{center}{whether to center the columns}

\item{scale}{whether to standardize the columns}
}
\description{
genpca
}
\examples{

coords <- expand.grid(x=seq(1,100), y=seq(1,100))
img <- apply(coords, 1, function(x) {
  x1 <- 1 - pnorm(abs(x[1] - 50), sd=8)
  x2 <- 1 - pnorm(abs(x[2] - 50), sd=8)
  x1*x2
})

mat <- matrix(img, 100,100)
mlist <- replicate(50, as.vector(mat + rnorm(length(mat))*.8), simplify=FALSE)
X <- do.call(rbind, mlist)
S <- neighborweights:::spatial_smoother(coords, sigma=8, nnk=27)

gp1 <- genpca(X, A=S, ncomp=2)

Xs <- do.call(rbind, lapply(1:nrow(X), function(i) X[i,,drop=FALSE] \%*\% S))
gp2 <- genpca(as.matrix(Xs), ncomp=2, center=FALSE)
}
