% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpca.R
\name{genpca}
\alias{genpca}
\title{genpca}
\usage{
genpca(X, A, M, ncomp = min(dim(X)), center = FALSE, scale = FALSE)
}
\arguments{
\item{X}{the data matrix}

\item{A}{the column constraints. Can be a \code{vector}, \code{matrix}, or sparse matrix.}

\item{M}{the row constraints. Can be a \code{vector}, \code{matrix}, or sparse matrix.}

\item{ncomp}{the number of components to estimate}
}
\description{
genpca
}

