% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpca.R
\name{genpca}
\alias{genpca}
\title{genpca}
\usage{
genpca(X, A = rep(1, ncomp(X)), M = rep(1, nrow(X)), ncomp = min(dim(X)),
  center = TRUE, scale = FALSE)
}
\arguments{
\item{X}{the data matrix}

\item{A}{the column constraints. Can be a \code{vector}, \code{matrix}, or sparse matrix.}

\item{M}{the row constraints. Can be a \code{vector}, \code{matrix}, or sparse matrix.}

\item{ncomp}{the number of components to estimate}

\item{center}{whether to center the columns}

\item{scale}{whether to standardize the columns}
}
\description{
genpca
}
