% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfa.R
\name{mfa}
\alias{mfa}
\title{multiple factor analysis}
\usage{
mfa(X, ncomp = 2, preproc = center(), normalization = c("MFA", "RV",
  "None", "RV-MFA", "Dual-RV", "Frob", "custom"), M = NULL, A = NULL,
  ...)
}
\arguments{
\item{X}{a \code{block_matrix} object}

\item{ncomp}{the number of components to estimate}

\item{preproc}{a preprocessing pipeline, default is `center()`}

\item{normalization}{the normalization method: MFA, RV, RV-MFA, or None (see details).}

\item{M}{custom weight matrix for the rows}

\item{A}{custom weight matrix for the columns}
}
\description{
mfa
}
\examples{

X <- block_matrix(replicate(5, { matrix(rnorm(10*10), 10, 10) }, simplify=FALSE))
res <- mfa(X, ncomp=3, normalization="MFA")
p <- partial_scores(res, 1)
stopifnot(ncol(scores(res)) == 3)

labs <- letters[1:10]
cfier <- classifier(res, labels=labs, ncomp=3)
pred <- predict(cfier, X[1:2,])
cfier2 <- classifier(res, labels=labs, ncomp=3, colind=res$block_indices[[2]])
pred2 <- predict(cfier2, X[1:2,res$block_indices[[2]]])
}
