% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mubada.R
\name{mubada}
\alias{mubada}
\title{Multiple Subjects Barycentric Discriminant Analysis}
\usage{
mubada(Y, Xlist, ncomp = 2, center = TRUE, scale = FALSE,
  normalization = c("MFA", "RV", "None", "RV-MFA", "custom"), A = NULL)
}
\arguments{
\item{Y}{dependent \code{factor} variable. If All X matrices have same number of rows, Y can be a single factor.
If there are a different number of rows (e.g. different numbers of replications per subject), Y can be a list of factors.}

\item{Xlist}{a \code{list} of X matrices, one per subject, or it is a \code{list} of \code{projector} objects.}

\item{ncomp}{number of common components to estimate.}

\item{center}{whether to center the variables.}

\item{scale}{whether to scale the variables by 1/sd.}

\item{normalization}{the type of normalization.}

\item{A}{optional constraint matrix for the columns used when \code{normalization} is "custom".}
}
\description{
Multiple Subjects Barycentric Discriminant Analysis
}
\references{
Abdi, H., Williams, L. J., & Béra, M. (2017). Barycentric discriminant analysis. \emph{Encyclopedia of Social Network Analysis and Mining}, 1-20.

Abdi, H., Williams, L. J., Connolly, A. C., Gobbini, M. I., Dunlop, J. P., & Haxby, J. V. (2012). Multiple Subject Barycentric Discriminant Analysis (MUSUBADA): how to assign scans to categories without using spatial normalization. \emph{Computational and Mathematical Methods in Medicine}, 2012.
}
