% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asca.R
\name{musu_asca}
\alias{musu_asca}
\title{musu_asca}
\usage{
musu_asca(Xlist, formula, ncomp = 2, design, center = TRUE, scale = FALSE,
  svd.method = "fast")
}
\arguments{
\item{Xlist}{the list of data matrices}

\item{formula}{a formula specifying the ANOVA design}

\item{design}{a \code{data.frame} providing the variables provided in \code{formula} argument.}

\item{center}{whether to center the variables}

\item{scale}{whether to scale the variables}

\item{svd.method}{the svd method to use.}
}
\description{
musu_asca
}

