% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asca.R
\name{musu_asca}
\alias{musu_asca}
\title{musu_asca}
\usage{
musu_asca(Xlist, formula, design, center = TRUE, scale = FALSE,
  svd.method = "base", strata)
}
\arguments{
\item{Xlist}{the list of data matrices}

\item{formula}{a formula specifying the ANOVA design}

\item{design}{a \code{data.frame} providing the variables provided in \code{formula} argument.}

\item{svd.method}{the svd method to use.}

\item{strata}{name of variable in design matrix that indicates any blocking structure, such as "subject".}
}
\description{
musu_asca
}
