% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuroca.R
\docType{package}
\name{neuroca}
\alias{neuroca}
\alias{neuroca-package}
\title{neuroca - multiblock component analysis methods for neuroimaging data}
\description{
neuroca - multiblock component analysis methods for neuroimaging data
}
\author{
Bradley Buchsbaum <brad.buchsbaum@gmail.com>
}
