% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepper.R
\name{prep}
\alias{prep}
\title{prepare a dataset by applying a pre-processing pipeline}
\usage{
prep(x, ...)
}
\arguments{
\item{...}{extra args}

\item{the}{pipeline}
}
\description{
prepare a dataset by applying a pre-processing pipeline
}
