% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{reconstruct}
\alias{reconstruct}
\title{reconstruct the original data with some number of components}
\usage{
reconstruct(x, ncomp)
}
\arguments{
\item{x}{the model fit}

\item{ncomp}{the number of components to use}
}
\description{
reconstruct the original data with some number of components
}

