% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asca.R
\name{residualize}
\alias{residualize}
\title{residualize}
\usage{
residualize(form, X, design)
}
\arguments{
\item{form}{the formula defining the model to fit for residuals}

\item{X}{the response matrix}

\item{design}{the \code{data.frame} containing the design variables specified in \code{form} argument.}
}
\description{
residualize
}

